refactor: separate cli.ts into modular components
Split cli.ts responsibilities into dedicated modules following SRP:
- cli/types.ts: shared type definitions and zod schemas
- cli/api-client.ts: API communication (fetchCompanies)
- cli/oauth-flow.ts: OAuth authentication flow
- cli/configuration.ts: settings persistence (saveConfig)
- cli/prompts.ts: UI/prompt handling with inquirer
- cli/index.ts: entry point exporting configure function

The original cli.ts now re-exports from cli/index.ts for
backward compatibility.

Closes #158
